support export after save model (#13844)
